requirements.txt
streamlit
PyPDF2
langchain
openai
langchain_openai
langchain_core
faiss-cpu
pytesseract
langchain_community
scikit-learn
PyYAML
streamlit
streamlit-authenticator
bcrypt
pillow
requests
numpy
chromadb
schema
pymongo
dnspython
chardet
wikipedia
streamlit_option_menu 
beautifulsoup4
streamlit_cognito_auth
boto3
CognitoAuth
